serving/serving_api/requirements.txt
python-dotenv
flask
elasticsearch